flask/requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
distlib==0.3.5
dnspython==2.2.1
filelock==3.8.0
flake8==5.0.4
Flask==2.2.2
Flask-Compress==1.12
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-swagger==0.2.14
flask-talisman==1.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
identify==2.5.3
idna==3.3
imagekitio==2.2.8
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.17.0
mccabe==0.7.0
nodeenv==1.7.0
numpy
packaging==21.3
pandas
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
redis==4.4.0rc1
requests==2.28.1
sentry-sdk==1.9.4
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.11
virtualenv==20.16.3
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
